requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
djangorestframework==3.15.1